added signup page

+++ Backend/routes/chat.js
@@ -24,9 +24,15 @@ router.post("/test", async(req, res) => {
 });
 
 //Get all threads
-router.get("/thread", async(req, res) => {
+router.get("/thread",auth, async(req, res) => {
     try {
-        const threads = await Thread.find({}).sort({updatedAt: -1});
+        console.log("=== GET THREADS ===");
+    console.log("req.user:", req.user);
+
+        const threads = await Thread.find({ userId: req.user.id });
+        console.log("threads found:", threads.length);
+    console.log("thread userIds:", threads.map(t => t.userId));
+
         //descending order of updatedAt...most recent data on top
         res.json(threads);
     } catch(err) {
@@ -35,11 +41,11 @@ router.get("/thread", async(req, res) => {
     }
 });
 
-router.get("/thread/:threadId", async(req, res) => {
+router.get("/thread/:threadId",auth, async(req, res) => {
     const {threadId} = req.params;
 
     try {
-        const thread = await Thread.findOne({threadId});
+        const thread = await Thread.findOne({ threadId, userId: req.user.id });
 
         if(!thread) {
             res.status(404).json({error: "Thread not found"});
@@ -52,11 +58,11 @@ router.get("/thread/:threadId", async(req, res) => {
     }
 });
 
-router.delete("/thread/:threadId", async (req, res) => {
+router.delete("/thread/:threadId",auth, async (req, res) => {
     const {threadId} = req.params;
 
     try {
-        const deletedThread = await Thread.findOneAndDelete({threadId});
+        const deletedThread = await Thread.findOneAndDelete({ threadId, userId: req.user.id });
 
         if(!deletedThread) {
             res.status(404).json({error: "Thread not found"});
@@ -75,19 +81,25 @@ router.post("/chat",auth, async(req, res) => {
     const userId = req.user.id;
 
     if(!threadId || !message) {
-        res.status(400).json({error: "missing required fields"});
+    return res.status(400).json({error: "missing required fields"});
     }
 
+
     try {
-        let thread = await Thread.findOne({threadId});
+        let thread = await Thread.findOne({ threadId, userId: req.user.id });
+
 
         if(!thread) {
             //create a new thread in Db
             thread = new Thread({
                 threadId,
                 title: message,
+                userId: req.user.id,
                 messages: [{role: "user", content: message}]
             });
+            console.log("=== CREATING THREAD ===");
+console.log("userId being saved:", req.user.id);
+
         } else {
             thread.messages.push({role: "user", content: message});
         }
